Move height mode related code
