+++ src/alignment/left.rs
@@ -7,7 +7,7 @@ use crate::{
         line::{StyledLineIterator, UniformSpaceConfig},
         StateFactory, StyledTextBoxIterator,
     },
-    style::HeightMode,
+    style::height_mode::HeightMode,
     utils::font_ext::FontExt,
     StyledTextBox,
 };
